Fixed bug
Added a check to the self global alignment to make sure the operons being compared for duplicates have been marked b/c it might caused an issue if you have two operons that are identical but have not been marked as orthologs in the previous steps. They'll reference each other as duplicates.